requirements.txt
# Core
flask==3.0.3
flask-socketio==5.3.6
gunicorn==22.0.0
eventlet==0.36.1
python-dotenv==1.0.1

# AI
openai==1.40.0
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
tiktoken==0.7.0

# Database
supabase==2.7.4

# Cache
redis==5.0.8

# WhatsApp
twilio==9.2.3

# Auth
PyJWT==2.9.0

# Export
reportlab==4.2.2

# HTTP
requests==2.32.3
httpx==0.27.0

# Monitoring
psutil==6.0.0
